Updated vector calculus tutorial
